For TestResolverShutdown, use address that can't be resolved.
This test only currently works because stun.l.google.com has an IPv4
address and the TURN port is created with an IPv6 address. But the test
would start failing if/when it starts providing an IPv6 address. Which
may already be happening, as indicated by a recent test failure.

Review URL: https://codereview.webrtc.org/1290233003