Ensure all test targets are built on Android.
When GYP runs for OS=android it doesn't generate the
video_engine_core_unittests_apk_target target which is needed to
get the APK built.
The same problem applies to webrtc/test/webrtc_test_common.gyp,
but that unittest is not added on any bot anyway, so that's future work.

TESTED=Ran webrtc/build/gyp_webrtc for Linux and Android locally.
Before this patch, the video_engine_core_unittests was not built
as part of the 'All' target. With this patch, it is now built.

R=pbos@webrtc.org

Review URL: https://codereview.webrtc.org/1348093002 .