Rewrote perform_action_on_all_files to be parallell.
This speeds up decode for a 720p workload from 25s to 5s on my machine.
I did some benchmarking and it appears we spend the vast majority of
the time on zxing decoding, so now we will bring all processors on the
the machine to bear.

BUG=webrtc:4986

Review URL: https://codereview.webrtc.org/1428433002