Flesh out webrtc/.gitignore
Chromium only checks out the webrtc directory so it misses the root
level .gitignore file which leads to messy "git status" reports inside
third_party/webrtc. This copies the root level .gitignore so that
.vcxproj files, the OSX equivalent, and other files will be ignored.

Some of the entries are irrelevant, but it is better too have a few
irrelevant entries than to be missing some, and the simplicity of
copying is valuable.

NOTRY=True
TBR=henrika@webrtc.org

Review URL: https://codereview.webrtc.org/1432413002