Add ApplyPacketOptions()
When libjingle is compied with ENABLE_EXTERNAL_AUTH the sending socket
needs to update RTP header in order for the outgoing packet to be
valid. The corresponding code was in chromium in
content/browser/renderer_host/p2p/socket_host.cc and it was impossible
to reuse it anywhere else. This CL moves this code to
talk/media/base/rtputils.h/cc, so it can be used outside of chrome.

BUG=crbug.com/547158
R=pthatcher@webrtc.org

Review URL: https://codereview.webrtc.org/1578323002 .