Rename RWLockGeneric to RWLockWinXP to more accurately reflect when it's used.
Since this is on Windows only, I'm also using the CriticalSectionWrapper and ConditionVariableWrapper Windows types directly which allows us to skip 3 extra heap allocations. It also helps with the removal of the 'friend' relationship between ConditionVariableWrapper and CriticalSectionWrapper, which is causing headaches on Mac.

BUG=

Review URL: https://codereview.webrtc.org/1595983002