PRESUBMIT: Exclude supplement.gypi from _CheckNoSourcesAboveGyp check.
The webrtc/supplement.gypi contains a source listing of the
suppressions for TSan and LSan, which are located in source files.
It makes the presubmit fail when it's updated, which is confusing.

NOTRY=True
TESTED=Edited webrtc/supplement.gypi and ran "git cl presubmit" without an error.

Review URL: https://codereview.webrtc.org/1649423002