NetEq: Fix a negative shift value
In some rare occations (very low energy signal), a shift value happened
to be negative. This is now fixed by using the WEBRTC_SPL_SHIFT_W32,
which in essence checks the sign of the number of shifts and performs a
right or left shift accordingly.

The fix reverts to how the code was written in old NetEq; see
https://chromium.googlesource.com/external/webrtc/+/4d363ae305fa975e7f3ad2a5ca5dbe5e3f101f82/webrtc/modules/audio_coding/neteq/normal.c#165.

BUG=webrtc:5490

Review URL: https://codereview.webrtc.org/1675293002