Add tools/mb to setup_links.py
I believe this is needed in order to setup the iOS simulator bots
using the ios recipe module. We'll also most likely use MB in the future
when moving over from GYP to GN (since that's the sole purpose of the tool:
to make such a migration easier).

BUG=chromium:498746
NOTRY=True
TBR=phoglund@webrtc.org

Review URL: https://codereview.webrtc.org/1692543002