Android: Add more info for createPbufferSurface() exceptions
This CL adds the width and height to the createPbufferSurface exception
message. Also, when a Callable.call() fails in
ThreadUtils.invokeUninterruptibly() we rethrow a new exception, but that
excludes the callstack from Callable.call(). This CL adds the callstack
from Callable.call() to make debugging easier.

BUG=b/27581640,b/27516991

Review URL: https://codereview.webrtc.org/1780183005