Delete unused video capture support for cropping, non-square pixels, and ARGB screencast scaling.
First two are unused, because the instance variables ratio_w_,
ratio_h_, and square_pixel_aspect_ratio_, are never modified after
initialization to 0 and false.

ARGB is believed to be unused, and the scaling logic
is probably not appropriate in any case.

Also delete corresponding helper functions in
videocommon.cc.
BUG=webrtc:5682

Review-Url: https://codereview.webrtc.org/1934503002