Enable muted state by default in VoE
This change turns muted state on by default in VoiceEngine, but not
for NetEq or AudioCodingModule when used stand-alone.

The expected effect is that voice channels that have not received any
packets for some time should reduce their CPU usage. This should have
a noticeable effect on endpoints with many incoming streams, but where
only a few have packets incoming at any given time (i.e., where an
intermediate server filters out the majority of the streams).

BUG=webrtc:5606
NOTRY=True

Review-Url: https://codereview.webrtc.org/1987143003