Fix inconsistent setting of the rtc_include_unittests flag.
On advice from brettw, se discussion here:
https://codereview.webrtc.org/2149543002/

The problem was we were setting the flag to both false and true,
and the the true happened to win out for WebRTC checkouts and
false for Chromium checkouts. This change should make this
mechanic more obvious.

This change _should_ have no effect downstream.

Doing tbr to see if we can get the chromium import back into a good state today.

TBR=tommi@webrtc.org

Review-Url: https://codereview.webrtc.org/2154693002