Fix for data channels perpetually stuck in "closing" state.
If the data transport is destroyed while data is buffered (due to
the PC being closed, or a description set with data rejected), the
data channel was getting stuck in a "closing" state, waiting to
finish sending its buffered data. But since there's no more transport,
it will never get another chance to send buffered data.

It just needs to terminate non-gracefully and discard the buffered data
in this situation.

R=skvlad@webrtc.org, zhihuang@webrtc.org

Review URL: https://codereview.webrtc.org/2235843003 .