[GN] Add rtc_sdk_framework_objc target to GN
The build artifacts don't look completely identical to the ones generated
by the GYP targets, but manual review shows the same symbols are exported.

On iOS,  the version generated by the GN follows convention, including
a "Headers" directory, and the .modulemap file. I think this is preferred
over the gyp version.

BUG=webrtc:6320
NOTRY=True
TESTED=Run AppRTCDemo on iOS + Mac and verified with nm that they export the same symbols.

Review-Url: https://codereview.webrtc.org/2340633003