Improves resolution when logging rate in the ADB class.
Trivial patch which fixes an issue where logged rate estimates could be
invalid. E.g. on iOS, two successive timer interrupts can be ~10.5 seconds
and not exactly 10.0 (which is usually the case on Android). In those
cases we could log a rate estimate of e.g. ~51000Hz instead of ~48000Hz.

This CL fixes that.

BUG=NONE

Review-Url: https://codereview.webrtc.org/2350103002