Make 'webrtc' a static library.
Most other targets were converted to static_library to match the
GYP build in https://codereview.webrtc.org/2361623004.

The monolithic 'webrtc' wasn't converted due to a limitation in
libtool on Mac: it will fail if there are no object files to link
(see http://crbug.com/625874).
Because of that, a no_op_function.cc source file was added to
the sources here, even if it's not used. This will be further
refactored in the future as part of the Slim and Modular
effort (bugs.webrtc.org/5716).

BUG=webrtc:6418, webrtc:6410
NOTRY=True

Review-Url: https://codereview.webrtc.org/2372983002