Fixed flaky SharedExclusiveLock tests.
These tests were checking the behavior of thread synchronization
primitives using nothing but carefully timed sleeps. This was very
unreliable and prone to flakes.

This CL replaces the sleeps with waiting on synchronization events.
There is still the need to wait for timeouts when testing for negative
outcomes, but it's greatly reduced. I've run these tests for thousands
of iterations on MSan without a single failure.

BUG=webrtc:5824
R=pthatcher@webrtc.org

Review URL: https://codereview.webrtc.org/2393023002 .