Remove useless debugging code
It would be enough to say we're removing EnableSrtpDebugging because
it's never called, but the story is a bit more interesting.

libsrtp's debugging facilities are gated behind the reasonably-named
ENABLE_DEBUGGING macro:

https://chromium.googlesource.com/chromium/deps/libsrtp/+/b17c065a8a63725eca216fd461984f82c331dd07/srtp/crypto/include/err.h#186

This code was imported to WebRTC from libjingle, but neither WebRTC or
Chromium ever set ENABLE_DEBUGGING. Even if someone had ever called
EnableSrtpDebugging, it wouldn't have done anything.

BUG=0

Review-Url: https://codereview.webrtc.org/2409513002