NetEq: Remove special case for Merge without Expand
(This is a backport of revision 14627 (c9ec8758db79c6fe4de). Original
CL here: https://codereview.webrtc.org/2412883002)

This was an ill tested special case which turned out to be more problem
than benefit. The special case was only triggered when the decoder frame
size was smaller than 10 ms, which is more or less unsupported by NetEq.

Also fixed a bug in a test, a bug which was exposed by the code change.

BUG=chromium:654983, chromium:653656
R=ossu@webrtc.org

Review URL: https://codereview.webrtc.org/2445373002 .