Fixed a bug where only the tests in the first shard were run.
This is because:
1) The environment variables were still around when the test was executed.
2) gtest-parallel executes only one test at a time.

So when the test was invoked from a shard different than the 0th one, for example as:
./something_unittests --gtest_filter=Test.Name
It read the environment variables, and the environment variables together with the gtest_filter flag, told it that there were several shards, but only one test to be run, so if it wasn't the 0th shard, it wouldn't run the test at all.

This is fixed by erasing the environment variables once read.

Also change swarming-related flag names to fit the rest of the flags and validate their values.

NOTRY=True
BUG=chromium:497757, chromium:664425
TBR=pbos@webrtc.org, kjellander@webrtc.org

Review-Url: https://codereview.webrtc.org/2505093003