Create top-level dir tools-webrtc and start moving things into it.
In order to get rid of the Chromium checkout for WebRTC, the plan is to
instead of cloning all of Chromium, only clone the build, third_party and
tools sub-directories. In order to do so, we must first move all things
checked into the WebRTC tools/ directory somewhere else.
Due to many hardcoded assumptions of tools/ existing in Chrome, this
is only manageble solution to the problem.

This first step only moves stuff not used by the build system or bots,
and deletes a few unused directories.

BUG=webrtc:5006
R=henrika@webrtc.org

Review-Url: https://codereview.webrtc.org/2584433002 .