Merge to M57: Only set certificate on DTLS transport if fingerprint is found in SDP.
This is used for fallback from DTLS to SDES encryption, which we probably still
want to support. Setting a certificate puts the DTLS transport in a "DTLS
enabled" mode, so it should be delayed until SDP with "a=fingerprint" is set.

BUG=chromium:690069
NOTRY=True
NOPRESUBMIT=True

Review-Url: https://codereview.webrtc.org/2681403002
Cr-Commit-Position: refs/branch-heads/57@{#4}
Cr-Branched-From: e5cbc2019003dbb40e03811d7607feb95757a4ec-refs/heads/master@{#16123}