iSAC float decoder: Don't read past end of initialized part of buffer
This is a backport to M57 of this CL: https://codereview.webrtc.org/2659383002

We read past the end of the initialized part of the buffer, seemingly
on purpose (no one knows the details of this code anymore). The right
thing to do is probably to zero that part of the buffer.

(The *right* right thing to would be to rewrite this so that it was
easier to see what data was supposed to be where when, but
priorities...)

NOTRY=true
NOPRESUBMIT=true
BUG=chromium:683040

Review-Url: https://codereview.webrtc.org/2697683003
Cr-Commit-Position: refs/branch-heads/57@{#5}
Cr-Branched-From: e5cbc2019003dbb40e03811d7607feb95757a4ec-refs/heads/master@{#16123}