Make lifetime of SimulcastEncoderAdapter implementation name safe.
Before this CL, the |codec_name| in the CodecSpecificInfo in the
OnEncodedImage callback chain depended on |implementation_name_|
which could be modified on a different thread.

This CL moves the ownership of the string that |codec_name| points
to into the callback adapter, thus making access to it safe.

Problem initially spotted by pbos@ in
https://bugs.chromium.org/p/webrtc/issues/detail?id=7314.

BUG=webrtc:7314

Change-Id: Ifefb9cfc8188319b2c2c5aba594e9573a6d98b5a