Remove deprecated AudioMixerImpl creation method.
AudioMixerImpl::CreateWithOutputRateCalculator has become
deprecated. Instead, either Create() or Create(OutputRateCalculator,
bool use_limiter) should be used. The first uses sensible default
values for missing arguments. The second takes all arguments. The old
CreateWithOutputRateCalculator is deprecated so that we don't have
different Create:s with all possible combinations of parameters.

Note that the factory methods may change in the future. The reason for
adding 'use_limiter' was that the limiter that was used had
questionable benefit and was very computationally expensive. Now work
is going on in WebRTC to replace it with a much cheaper version. After
the change, the factory method may change again to not allow for
disabling the limiter.

Bug: webrtc:7167
Change-Id: I0f9005e27e726fa552ee38dcbe965274e5006544