Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I had forgotten about the problem with different supported resolutions for the different cameras. If we take the union of all resolutions, what happens when we select a resolution that is not supported by the front camera, will AppRTCMobile even start?

It starts fine. In the file `ARDCaptureController.m` the method `-selectFormatForDevice:` seems to find the closest matching supported video resolution to the one stored in settings. So it should match exactly for at least one of the cameras, and if not it will choose the closest supported format. It is the only place in the code I can find that uses the user setting for video resolution.

Patch-set: 2
Reviewer: Anders Carlsson <1215177@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>