Update patch set 5
Patch Set 5:

After thinking about this a bit more I think there is a problem with this approach, namely the case where NetworkTraversal returns v4 and v6 URIs. If we replace both of those with a hostname URI, only one of the v4/v6 IPs can be written to the IceServer.ip field.

We could probably come up with a workaround for this, but things are starting to get kind of complicated, and Emad's previous approach might end up being the simplest.

Taylor, wdyt?

Patch-set: 5
Reviewer: Justin Uberti <1126895@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>