WIP Hook up HardwareVideoDecoder
Create VideoDecoderFactory interface and implementation.

The implementation creates an Android hardware video decoder.  It is built
around the same patterns as the HardwareVideoEncoderFactory.

This change pulls some shared code and constants into a common "utils" class.

Finally, adds an instrumentation test for the HardwareVideoDecoder.

BUG=webrtc:7760

Change-Id: Iea6eaae7727925743cb54f7c3153a6c07d62f55d

Create a hardware VideoDecoder implementation using Android MediaCodec.

BUG=webrtc:7760

Change-Id: Ieae3852d22cadf24cf4184ae985062918a85f02c

Create the VideoEncoderFactory and implement it.

Adds the VideoEncoderFactory interface and implements it for use with HardwareVideoEncoder.  This uses MediaCodecVideoEncoder's initialization code as an example.

BUG=webrtc:7760

Change-Id: I9fbc93ce9ac4ad866750a4386c4f15e800a3073e