Update patch set 16
Patch Set 16:

> Patch Set 16:
> 
> OK @pbos @sprang, got back to this and:
> 
> 1. Found and fixed an illegal stack access that was causing the valgrind failure and a bunch of other segfaults on some other builders.
> 
> 2. Increased a timeout in some of the end-to-end tests since the timeout was too short when running under valgrind.
> 
> 3. Rebased: There were merge conflicts from some commits that have gone into master since.
> 
> Ran unittests with valgrind on Linux and on Mac without. Could we try another CQ attempt now?

The increased timeout in the simulcast end to end test am referring to is here:

https://chromium-review.googlesource.com/c/514351/16/webrtc/video/end_to_end_tests.cc#4024

The unit tests pass with the previous 30s timeout when run directly but fail when running under valgrind. Increasing the timeout to 120s fixes the issue.
I considered adding a valgrind specific constant for the timeout, but ended up directly increasing it in all cases instead since it is only ever used in the test as a maximum wait time in an rtc::Event.

Patch-set: 16
Reviewer: Hemanth Meenakshisundaram <1176415@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>