Only create H264 frames if there are no gaps in the packet sequence number.
In the case of H264 we can't know which packet that is the fist packet of a
frame. In order to avoid creating incomplete frames we keep track of which
packets that we haven't received, and if there is a gap in the packet sequence
number leading up to this frame then a frame wont be created.

BUG=chromium:716558
R=holmer@google.com, stefan@webrtc.org

Review-Url: https://codereview.webrtc.org/2926083002
Cr-Original-Commit-Position: refs/heads/master@{#18559}
Review-Url: https://codereview.webrtc.org/2947373002 .
Cr-Commit-Position: refs/branch-heads/60@{#5}
Cr-Branched-From: c61bf947b4ac31f3500858ffcae6fee39d799930-refs/heads/master@{#18252}