Update patch set 2
Patch Set 2:

I am simply trying something out to ensure that the tests "survives" on this particular bot. I can run the tests 100 times on different devices and it works well. I would like to keep it as is but this is just an attempt to avoid disabling the test totally. To summarize: it is better than nothing. If it does not work, I might consider adding a sleep as next step and disable. The exact sequence (Start/Stop/Start/Stop) is rare and not a typical use case.

Patch-set: 2
Reviewer: Henrik Andreasson <1001300@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0