Merge of Corrected the computation of the headroom in the AEC3 buffer alignment
This is a merge of the CL
https://chromium-review.googlesource.com/549335 in the M60

This CL corrects the computation of the delay headroom so that
it is only updated when the delay is updated. This is important
as otherwise a too large headroom will be reported, which then
could cause buffer access issues.

TBR=peah@webrtc.org

Bug: webrtc:7878, chromium:736893
Change-Id: Ib37cb608b064dd5d4df3f8fc423448ee80ed0106
Reviewed-on: https://chromium-review.googlesource.com/549335
Commit-Queue: Per Åhgren <peah@webrtc.org>
Reviewed-by: Ivo Creusen <ivoc@webrtc.org>
Cr-Commit-Position: refs/heads/master@{#18781}
(cherry picked from commit f0a6fb19c6e8a7fc7c3da6be7b268bb376d33701)