Merge of Corrected the initialization of the AEC3
This CL is a merge of the CL  https://chromium-review.googlesource.com/549315
into M60.

This CL corrects the initialization of the AEC3, as well
as for the other submodules in the whole audio processing module
in the sense that it properly update the submodule states also
for the case when reinitialization is trigger from the render
side of the audio processing module.

NOTRY=true
NOPRESUBMIT=true

Bug: chromium:736889,webrtc:7879
Change-Id: I423e963835d0c3227caa8e186b29031bcb912515
Reviewed-on: https://chromium-review.googlesource.c om/549315
Commit-Queue: Per Åhgren <peah@webrtc.org>
Reviewed-by: Ivo Creusen <ivoc@webrtc.org>
Cr-Original-Commit-Position: refs/heads/master@{#18784}
Review-Url: https://codereview.webrtc.org/2965953002
Cr-Commit-Position: refs/branch-heads/60@{#8}
Cr-Branched-From: c61bf947b4ac31f3500858ffcae6fee39d799930-refs/heads/master@{#18252}