Update patch set 2
Patch Set 2:

Using the current time as a seed instead of std::random_device() prevents crashing when it cannot open /dev/urandom. SystemTimeNanos() is used instead of TimeNanos() so that it does not pass in 0 when a fake clock is used in tests (rtc::Random fails if the seed is 0).

Patch-set: 2
Reviewer: Steve Anton <1228509@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>