Implementation of SSL caching/resumption, with unit tests. Minimal version in https://chromium-review.googlesource.com/c/575910/.
This CL adds the ability for a SSLAdapter to resume a previous session, saving a roundtrip and significantly reducing the # of bytes needed to bring up the new session.

To do this, the sessions need to share state. This is addressed by introducing the SSLAdapterFactory object, which can maintain a SSL_CTX and session cache for multiple sessions.

This CL also includes unit tests, which verify resumption through the new IsResumedSession method. These tests required some nontrivial pounding on SSLStreamAdapter, where a similar factory object was introduced in order to allow the SSL server to share a SSL_CTX.

Bug:7936
Change-Id: I25cf4945b5beaa92568b63c816b6bfae32b9567b