Update patch set 3
Patch Set 3:

I'm not sure this approach will work properly. Currently MouseCursorMonitor reports mouse position only when the mouse is on top of the selected window. If the selected window is covered by another window and the mouse cursor is on top of that window then cursor is considered to be outside of the selected window. I don't think you can keep this behavior if the capture monitor is not aware of the selected window.

Patch-set: 3
Reviewer: Sergey Ulanov <1116023@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>