Update patch set 7
Patch Set 7:

Thanks as always for the comments. I've added a few more tests to cover the scenarios we have been discussing and ensure the implementation matches what we think it should be.

I've rebased this on top of a new CL for renaming RsidResolutionObserver and will work on adding the additional hooks to this CL tomorrow.

There's still a few comments I haven't addressed yet, and I'll be getting to those soon.

And of course more tests will be coming tomorrow(TM).

Patch-set: 7
Reviewer: Steve Anton <1228509@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>