This CL adds the Total Harmonic Distorsion plus noise (THD+n) score in APM-QA.
In order to compute a THD score, a pure tone must be used as input signal.
Also, its frequency must be known. For this reason, this CL adds a number of
changes in the APM-QA pipeline. More in detail, input signal metadata is loaded
and passed to the THD evaluation score instance. This makes the eval_scores
module less reusable, but it is fine since the module has been specifically
designed for the APM-QA module.

Bug: webrtc:7494
Change-Id: Ie4f25b8e92bcb85d25eac9e1cb0c902eb373bd16