Update patch set 22
Patch Set 22:

Some responses to the open issues:

> Patch Set 18:
> 
> (3 comments)
> 
> Sorry for taking long to lgtm-ize. I think there remain three open issues, except for which I'd be happy to also +1 this CL:
> 1. The MID-RSID issue (comments inline).

If the MID is on the packet, but is unknown, we should drop the packet.  If the MID is known, but only in a (MID, RSID) pair and the packet does not match any (MID, RSID) pair, then we have something outside of the BUNDLE spec.  In that case, I think we should treat it as though we route to a track-level sink by MID that then drops it because it doesn't know which encoding it is.  In other words drop the packet, but still latch the SSRC to the MID (unlike when we drop for an unknown MID, in which case we don't latch the SSRC to the MID).  

> 2. The RSID/RRID issue - you've agreed inline that the behavior is not well-defined; could you please ping Taylor and Peter for resolution?

If a packet has both an RRID and RSID, then it must be a repair packet and only the RRID is of interest to us.  An RSID of a repair packet is not interesting, and I doubt it will ever be used.  It only exists in the spec as a possibility for historical and political reasons in the IETF.  It is of no practical use. 

> 3. Changing the unit-tests to match the changes that (I think) we agree on, from the RtcpDemuxer-CL. I'll review after that, but I don't think it should block.

Patch-set: 22
Reviewer: Peter Thatcher <1125619@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>