Split LogRtpHeader and LogRtcpPacket into separate versions for incoming and outgoing packets.
This is done in anticipation of proto format changes that will use different types for incoming and outgoing packets. This CL makes the interface more natural for the new event log implementation, and it is just as natural for the current implementation. (Note that the current format actually has fields that are only used for outgoing packets. However, we're not changing the encoding or implementation in this CL, merely adding a thin wrapper around the old code.)

BUG=webrtc:8111

Change-Id: I9ef2639761d775976d207f8c6f6d0a29f80ab52f