Move log_count_ out of RtcEventLogImpl
The limit on logs is not specific to the implementation of the logs, but is rather shared between all possible logs.
Also, by making it local to the .cc, not a member, we reduce the necessity of making RtcEventLog::Create a friend of the implementation. This necessity is removed completely by a following CL.

BUG=webrtc:8111
NOPRESUBMIT=True

Change-Id: I03044ed55ceeaf0064d5207b7407926571590699