Update patch set 5
Patch Set 5:

> Patch Set 5: Code-Review+1
> 
> No comments about the code, but two general questions. 
> 
> RenderDelayController now stores kMaxApiCallsJitterBlocks of the capture history in a circular buffer. To me, it seems a common enough thing to do that we should have a utility class for that. Is RenderDelayBuffer appropriate, can it be adapted or do we have something else? I think the code is good as is, this is more about potential areas of future improvement.
> 
> Delaying the capture buffer for kMaxApiCallsJitterBlocks before estimating delay has consequences: If the delay is too long, the matched filter will not have the render history that is far enough back in time that it matches the delayed capture. To keep the algorithm detecting as big delays, shouldn't the matched filter increase by the same number of samples?
> 
> Further, if the capture seen by the delay estimator is kMaxApiCallsJitterBlocks samples old, it will take as much time more until the delay estimator is confident of a new delay. How does that affect convergence speed and the echo canceller?
> 
> Of course, I don't know the trade offs, may have misunderstood, and if you think this is a net positive improvement, I don't object. The code LG and since I don't have any reason to not think it's a net positive improvement the change LGTM.

Thanks for the quick review!

Your comments are very valid, and I agree.

Regarding the circular buffer:
I think you are correct in that we should have a utility class for it. terelius@ has a proposal for that that we should look further at.

Regarding the delay:
You are correct. However, I've never seen delays that span the whole look window, so I think it is ok to reduce the look window in time. Furthermore, this functionality is a bit expensive so I'm hesitant to increase the length further. Let's look into that further though,

Regarding the delay in detection:
That is definitely correct. However, the delay is not that long (about 120 ms) which is fairly small compared to the time it takes to detect a delay change so it is ok. But we probably should change it to a solution where we instead shift the render buffer.

Patch-set: 5
Reviewer: Per Åhgren <5125@58829da1-049c-39fb-b951-ebdcd0984767>