Update patch set 1
Patch Set 1:

This is an exploratory cl to see what it takes to use the RtpPacketSinkInterface for receive stats.

We should move receive statistics to the RtpStreamReceiverController, I think, but do you see any easy first step? One problem is that RtpReceiver::IncomingRtpPacket wants an |in_order| argument, which is derived by the ReceiveStatistics objects. To clean this up, we'd need to either stop using that argument, or add a corresponding flag to RtpPacketReceived and set it close to where packets are parsed.

Which reminds me of another sub project: Moving parsing from Call to RtpStreamReceiverController, which requires moving configuration of header extensions. Maybe we have to do that first.

Patch-set: 1
Reviewer: Danil Chapovalov <5019@58829da1-049c-39fb-b951-ebdcd0984767>
Reviewer: Niels Moller <5234@58829da1-049c-39fb-b951-ebdcd0984767>
Work-in-progress: false